Changed floats to int and tweaked program flow.
Changed float arithmetic to integers (mostly uint16_t). This is to make
frequent calculations a bit faster. Also, made some functions static to
limit how the logic is accessed. Meaning that the main program only
needs to call `monitor_encoder()` to update check for obstructions and
velocity recording is done in the background.